requirements.txt
Flask
opencv-python
numpy
python-dotenv
cloudinary==1.36.0
flask-cors
gunicorn==21.2.0
flasgger
tensorflow==2.15.0
tensorboard==2.15.0
matplotlib==3.8.2